% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScDeconvExperiment.R
\docType{class}
\name{ScDeconvExperiment-class}
\alias{ScDeconvExperiment-class}
\title{ScDeconvExperiment class}
\description{
S4 class extending SingleCellExperiment with analysis parameters
and workflow history for single cell RNA-seq clustering and
deconvolution-based annotation.
}
\section{Slots}{

\describe{
\item{\code{params}}{A list of analysis parameters.}

\item{\code{history}}{A character vector describing performed analysis steps.}
}}

